refactor(config): replace config imports with clientConfig for client-side usage
The config file has been split into clientConfig to ensure client-side environment variables are used correctly. This change avoids potential issues with server-side environment variables being exposed to the client.
