+++ src/components/DiscordConnection/PrivateProfile.tsx
@@ -13,7 +13,7 @@ import { deleteUser, addNick, getUser } from "../../functions/requests/users";
 import { leaveClan } from "../../functions/requests/clans";
 import { supportedLanguages } from "../../config/languages";
 import { closeSession, getStoredItem } from "../../functions/services";
-import { DEFAULT_LANGUAGE } from "../../config/config";
+import { DEFAULT_LANGUAGE } from "../../config/clientConfig";
 import type { UserInfo } from "../../types/dto/users";
 
 const PrivateProfile = () => {
