+++ src/functions/requests/clans/members.ts
@@ -1,5 +1,5 @@
 import { getStoredItem } from "../../services";
-import { config } from "../../../config/config";
+import { config } from "../../../config/clientConfig";
 import type {
   MemberAction,
   MemberInfo,
