+++ src/functions/requests/maps/index.ts
@@ -1,4 +1,4 @@
-import { config } from "../../../config/config";
+import { config } from "../../../config/clientConfig";
 import type { GenericResponse } from "../../../types/dto/generic";
 import type {
   AddMapRequestParams,
